Pass shadowed action environment to the Starlark action.
This CL adds a new method to Action interface: `getEffectiveEnvironment` which returns the effective environment used to run the action based on given client environment. This method is called by the Starlark action to extend its environment with its shadowed action's environment before execution. The two environments variables are merged so that the Starlark variables can overwrite any of the shadowed action's environment variables.

PiperOrigin-RevId: 363384154