LocalFilesArtifactUploader skips isDirectory checks for non-directory artifacts.
When building with --nobuild_event_binary_file_path_conversion, BEP will
contain local URIs for artifacts. This is used by local tools (eg. intellij
plugin, IJwB) that execute a build then immediately read the built artifacts.
When combined with aspects, tools may use a build to extract arbitrary data
from the build graph.

Previously every artifact whose URI is included in BEP had its file type
checked so directories would not be erroneously included as files, and this
check issued a stat syscall. Often this check is unnecessary as the artifact is
produced by an action that has pre-declared the artifact as a file or
directory. By keeping this information during BEP production we can avoid stat
calls.

For an IJwB build with 9999 targets and 300,000 files in the BEP, this change
improved wall-clock time from 40s to 10s.

RELNOTES: None.
PiperOrigin-RevId: 363660858