Embed `PathFragment` in `Path`.
Number of `Path` methods overlap with `PathFragment` and given since
https://github.com/bazelbuild/bazel/commit/8883c615a47148e9fd14179cb8f20a7329c5c307, we always use `PathFragment` when performing actual IO (the only
case when holding onto a `Path` instead of `PathFragment` is beneficial), we
can include a `PathFragment` object in `Path` instead of duplicating the
members.

This change increases memory usage since every `Path` object now has an extra
pointer (to the `PathFragment`), but reduces the amount of garbage generated
when running IO operations and simplifies the code.

PiperOrigin-RevId: 363667423