Add ability to gate builtins injection on a per-symbol basis
See description of this behavior in the documentation of the new flag, --experimental_builtins_injection_override. The flag is allowMultiple, so it can be used in multiple places, but in practice it should only be set by a rules migrator working on coordinating the release of the Starlark implementation of a native ruleset.

Work toward #11437.

PiperOrigin-RevId: 363679375