Update BEP docs to explain aspect output.
Aspect results are available in BEP but do not have dedicated
events. Instead, the existing TargetConfigured and TargetComplete
events are reused for aspect results. Writing a tool to consume
BEP must be done carefully in the presence of aspects, so the
docs now explain these details, provide an example and warns
against common pitfalls.

PiperOrigin-RevId: 363737222