Update BEP docs to explain NamedSetOfFiles consumption; add example.
The NamedSetOfFiles event in BEP is tricky to consume due to its recursive,
shared structure and its ordering properties within the BEP stream. Consuming
these events is a common use case as it is necessary to obtain the output files
from a build. This change adds documentation specific to consuming these events
and provides sample pseudocode to that effect.

PiperOrigin-RevId: 363782854