Add runfiles.merge_all() for merging a sequence of runfiles objects.
We've seen some pathological cases where runfiles.merge() run in a loop results
in an extraordinarily deep and inefficient NestedSet representing the
runfiles's artifacts - and attempting to walk such a NestedSet could lead to a
stack overflow.

Merging a list of runfiles at once produces a NestedSet which is much shallower
(by two orders of magnitude in some cases).

In a followup to this change, we intend to introduce checks to make sure
NestedSet depth does not become excessive.

RELNOTES: Add runfiles.merge_all() for merging a sequence of runfiles objects.
PiperOrigin-RevId: 364399853