Fix bugs in applicable_licenses propagation so it goes to all parts of the build graph during analysis phase. This fixes the issue where an override of the attribute on a per-target basis wasn't being correctly applied in the graph. This also fixes issues where blaze query returned no information about license targets.
This requires that the old-style license checking path be aware of the
new-style license rules. Unfortunately, since the new-style license rules are
just plain old user defined rules, we must rely on matching the "magic" values
we picked for license.bzl in order to properly identify them. While this is
unfortunate, this hardcoding only occurs in parts of the old-style license
checking code that we ultimately will remove, so this hardcoding is by
definition short-lived and won't impact external users.

PiperOrigin-RevId: 364879893