Do not trim test configuration when --nodistinct_host_configuration is on.
Background is in b/117932061. The infrastructure avoids trimming the host configuration and thus the untrimmed target configuration is used in various places when --nodistinct_host_configuration is on. This leads to potential issues.

This may fixable more principally by allowing and ensuring the effective host configuration is trimmed when --nodistinct_host_configuration is in use; however, this requires far more intricate adjustments to configuration resolution.

PiperOrigin-RevId: 364898085