Emulate doing trimming at 'top-level' of AnalysisTestCase.
In real builds, the top-level targets (i.e. what is specified on the command line) will be analyzed under a properly trimmed configuration. e.g. if you do `bazel build :not-a-test-target --trim_test_configuration' then analysis of that target will not include TestOptions in the provided configuration.

AnalysisTestCase doesn't hook into the analysis phase early enough to trigger that sort of trimming. This CL adds the necessary logic to look for and trigger that trimming when needed. (Note that BuildViewTestCase already had to do similar!)

This indirectly fixes a handful of tests that break if --trim_test_configuration default is changed.

PiperOrigin-RevId: 366116560