Make shell tests less sensitive to --trim_test_configuration.
Change some shell tests to instead use --run_under or --platform_suffix as triggers to force a rebuild instead.

For build_event_stream_test, turn on --trim_test_configuration (explicit for now, can be removed when default is true) and adjust test logic to account for the new, trimmed configuration.

PiperOrigin-RevId: 366127923