Allow BuildViewTestCase to ignore some options when comparing configurations
Rather than simply comparing the two configuration, the test author can now include a set of options to 'exclude' from the comparison. They are trimmed away and resultant trimmed configurations are then compared.

Includes a related fix to GoConfiguredTargetTest to ignore TestOptions from the comparison (since the case where trim_test_configuration is on, the actual configuration will not include TestOptions and this is fine).

PiperOrigin-RevId: 366134162