Miscellaneous fixes to tests to avoid using --test_arg as no-op, etc.
Instead of using --test_arg as a no-op configuration setting, use a garbage cpu value (since not building anyway), define and define_values, or just change the associated targets to test targets (to avoid any trimming for when --trim_test_configuration is on).

These are not BuildViewTestCase so introducing the DummyTestFragment is non-trivial hence the use of 'different' 'no-op' configuration flags.

PiperOrigin-RevId: 366139254