Only accumulate necessary targets in `test_suite.$implicit_tests`
Avoids a test suite having a dependency on every test in the package when it
only wanted ones matching certain tags. This is also useful for not
accumulating a throwaway value of `$implicit_tests` when nothing needs it.

This does unfortunately duplicate some tag-based test filtering logic. I'm not
confident enough to go about dismantling all that, but I don't feel overly
guilty since this isn't the first time and it seems like this is the direction
test-suite filtering should go anyway.

This _might_ be considered an incompatible change in that it changes
`$implicit_test`'s value in the various query output formats, but I think the
wording of `test_suite.test`'s documentation is such that this could be
considered a bug fix or optimization.

PiperOrigin-RevId: 366246434