Further optimize same_pkg_direct_rdeps by not calling #getTarget on targets not in the same package.
`TargetProvider#getTarget` represents quite a significant chunk of overhead when loading additional packages. The previous implementation incurred this overhead when enqueuing the visit; this change shifts the overhead back to when the visit occurs.

For `same_pkg_direct_rdeps`, it visits the transitive closure of the package and stops when visiting targets from other packages. In this case, when visiting an external target (we can tell from just the `Label`), there is no need to load external packages (as we are only concerned about direct rdeps). Performance benefits vary based on how many external packages the package depends on, the more there are the greater the benefits.

PiperOrigin-RevId: 366292523