Adjust advice about propagating runfiles from attributes like srcs
Previously, this said to merge runfiles from attributes which provide source files. But that doesn't apply to targets that use source files in a way that's unrelated to those files' associated runtime dependencies.

For example, *_library.srcs provides source code which gets processed and eventually executed. And filegroup.srcs forwards its source files unchanged. In both of those cases, runfiles need to be merged from srcs.

But if a rule, for example, processes source code in some way that doesn't lead to the code's eventual execution, it may not make sense to merge runfiles associated with that code. If a target performs some sort of static analysis on source code in its srcs, something which needs the output of the analysis at runtime does not need data files which would be loaded by that code if it was executed.

This clarifies the wording by organizing it differently. Instead of grouping "data" with "deps" as "things which provide runtime dependencies", it puts "data" by itself (gets runfiles merged in by definition), and "srcs" and "deps" together (gets runfiles merged in if those might provide code for eventual execution).

RELNOTES: None.
PiperOrigin-RevId: 366299804