Query Docs: Make the section on "Words" tokens both correct and consistent, and also slightly improve readability while I am here.
The documentation been at least slightly wrong or slightly inconsistent for a long time. For example, commit 5e2d523 in Oct 2019 attempted to fix several inaccuracies (due to Bazel behavior changes making the docs stale) but in doing so accidentally introduced an inconsistency with respect to the period character.

Out of scope: Improving the documentation in other ways. That can come in the future :)

PiperOrigin-RevId: 366508062