Update documentation on //... and //:all
The intent of this patch is to be a little more explicit about these
two patterns. It appears that there currently isn't a spot that
explicitly documents these patterns.

I also decided to update the examples for "Target names" and "Package
names" because the `...` syntax does have a special meaning. I thought
it best to avoid confusion for those cases.

Closes #13171.

PiperOrigin-RevId: 366801719