Stop the Android dexer from using legacy toolchains with platforms.
This is way more complicated than you would expect.

Short story is the Android dexing aspect automatically propagates down the
":android_sdk" attribute, which references the legacy toolchain. This dexes needed
.jars in the toolchain's deps (and those deps can appear in any package, so
we can't easily identify them by filtering for the toolchain's package).

With platform-based toolchains, this breaks builds due to both types of
toolchains being read in different places and mixing poorly. See code comments
for details.

The only real solution is to *never* propagate down ":android_sdk"  with
platforms. This is also good general design: ":android_sdk" should have no
influence on platform-based Android builds under any circumstances.

But aspects have no mechanism to conditionally skip propagation. So this change
adds a small API extension to aspects to support that. It's a classic case of a
generic API to support an exceptional situation, so comments are also included
not to use it for other purposes.

PiperOrigin-RevId: 366890107