Normalize `--add-exports=` and `--add-opens=` flags
This is similar to the existing handling of `-source`, `-target`,
and `--release`, where (unlike javac) we allow all the options
to be mixed and matched, and let the last one win. With the
module-related flags, they are accumulated and added to the
command line only if the target level is 9 or higher, which
corresponds to modules being enabled.

PiperOrigin-RevId: 367471359