ConfigCommand should handle multiples of same key.
BuildConfigurationValue.Key actually is two things: a set of options and a set of what option classes the resultant BuildConfiguration should include. (The latter basically allows for another way to do trimming that we rarely exercise.) The ConfigCommand processing is using toString, which just looks at the checksum for the set of Options.

In the case of --trim_test_configuration, trimming is done via trimming the options but not the set of classes. Thus, the resultant key right after trimming has the same toString as the same configuration later on with the appropriately trimmed set of classes.

Handle this gracefully by merging the results into the same key-value pair for the processed map.

PS: This may fail (and likely would have before this CL) if the 'set of fragment classes' feature itself is being used to do some form of trimming.
PiperOrigin-RevId: 367593602