Simplify build failure output by always using `NNN arguments` (plural)
When Blaze reports a compilation error, it shows the first bit of the command line used to invoke the compiler. When more than 200 characters have been shown (`APPROXIMATE_MAXIMUM_MESSAGE_LENGTH`), the rest of the command line is shown as:
```
(remaining NNN argument(s) skipped)
```

When I see the message in my console, it's normally because 100+ arguments are being skipped. So the correct form is then:
```
(remaining NNN arguments skipped)
```

I find this more readable since it avoids a nested parenthesis and it uses the correct plural form. Put differently, the `argument(s)` form is asking the reader to do the mental work of selecting between `argument` and `arguments` — which is unnecessary in the vast majority of cases.

RELNOTES: Simplify build failure output by always using `NNN arguments`.
PiperOrigin-RevId: 367620498