Add data from JavaCcLinkParamsInfo provider to JavaInfo and add flags to disable the former.
All the rules that provide JavaCcLinkParamsInfo also provide JavaInfo. The data is added to JavaInfo in all such native rules.
Everywhere that JavaCcLinkParamsInfo is constructed, we first attempt to retrieve data the provider and if not available from JavaInfo.

Flags are added to disable the provider and to disable API related with the provider.
The CL is expected to introduce a minor regression, which shall be removed after the flags are flipped.

PiperOrigin-RevId: 367632113