Automated rollback of commit 52c0597dea3e2e620088c78a2f6a812d672c01dd.
*** Reason for rollback ***

Crash due to failed assertion, leading to issue 184944522.

*** Original change description ***

BEP reports analysis failure for targets with artifact conflicts.

Artifact conflicts are detected through a non-skyframe evaluation that is not
capable of identifying a top-level target or aspect that failed as a result of the
artifact conflict. We do perform a skyframe evaluation to identify which targets
and aspects to actually evaluate however, and *this* evaluation is now used to
report artifact conflicts in BEP.

RELNOTES: None.
PiperOrigin-RevId: 367682388