Automated rollback of commit bd165104ebb679f51fea084d92add1adf0091c06.
*** Reason for rollback ***

Breaks //storage/userdata/config/common:sdl_bundle_detection_test due to some genquery issues.

*** Original change description ***

Further optimize same_pkg_direct_rdeps by not calling #getTarget on targets not in the same package.

`TargetProvider#getTarget` represents quite a significant chunk of overhead when loading additional packages. The previous implementation incurred this overhead when enqueuing the visit; this change shifts the overhead back to when the visit occurs.

For `same_pkg_direct_rdeps`, it visits the transitive closure of the package and stops when visiting targets from other packages. In this case, whe...

***

PiperOrigin-RevId: 367699531