Move ObjcCppSemantics to cpp package
We would like to account for semantics in the build API.  This means
cpp package must be able to create ObjcCppSemantics.  But the objc
package depends on the cpp package so that would create a circular
dependency.  Break up this circular dependency by moving some objc
parts to cpp.

Note objc flags are exposed to bazel (there is no
GoogleObjcCppOptions), so we move the required flags to CppOptions.

PiperOrigin-RevId: 368061633