Handle genrule coverage configuration
srcs is treated as a source attribute because that's the predictable configuration for that attribute. (srcs is a source attribute for other native rules, with the one exception of filegroup, since that forwards inputs without processing.) That means coverage makes the general assumption:

* Files provided by targets in genrule.srcs are potentially coverage-relevant source code.

* The dependencies of stuff in genrule.srcs are not expected to be runtime dependencies.

This is not correct configuration for every use-case. For example, a genrule could take intermediate compilation products as input and produce binaries (or further intermediate products) as output. In that case, the transitive dependencies of genrule.srcs are coverage-relevant. However, in that case it would be preferable to use Starlark rules instead of genrule, with the intermediate outputs forwarded via rule-specific providers.

RELNOTES: Mark genrule.srcs as a source attribute for coverage.
PiperOrigin-RevId: 368238747