Propagate analysis-time aspect failures in AnalysisFailureInfo when using --allow_analysis_failures
When using --allow_analysis_failures, if an aspect implementation function
fails, bazel currently asserts because AspectFunction does not expect to get a
null ConfiguredAspect.

Instead, similarly to what we do for analysis-time failures in rules when using
--allow_analysis_failures, we want to create a special "errored aspect" and
propagate the failures in AnalysisFailureInfo. That way, the failures can be
tested for in bazel-skylib's analysistest with `expect_failure = True`.

RELNOTES: When using --allow_analysis_failures (for example, via bazel-skylib's
analysistest with `expect_failure = True`), analysis-time failures in aspect
implementation functions will now be propagated and saved in
AnalysisFailureInfo, just like analysis-time failures in rules.
PiperOrigin-RevId: 368504346