Fix resolved toolchains in cc rules not filtered by --noimplicit_deps
After https://github.com/bazelbuild/bazel/commit/31b0453b84eb2131ff075b9022cb2d6c95e1c71a, keys of resolved toolchains were already being added to implicit deps. However, the later comparisons to those keys were not properly accounting for the potential inclusion of toolchain context information.

ConfiguredTargetKey is the expected argument of ConfiguredTargetFunction which is named SkyFunctions.CONFIGURED_TARGET. Hence, the cast is safe. (Note that a literally dummy function and key using the same SkyFunction exists in NodeEntryPrinter; however, it would already be a major bug for that dummy key to end up here, in targetifyValues.)

Added tests so now there is a case using a mock rule, Starlark rule, and the cc rules (a native rule and toolchain).

PS:
Essentially, something in the cc rules implementation is taking the resolved cc toolchain and (accidentally?) registering it as a direct dependency. Notably this does not happen for Starlark-defined and used toolchains (but may happen with other native rules and toolchains).

Even if the registering is found in the cc rules, it would not be particularly incorrect. It remains peculiar that Starlark-defined rules and toolchains were already working and not exhibiting the key type discrepancy.

RELNOTES: cquery --noimplicit_deps now correctly filters out resolved cc_toolchains
PiperOrigin-RevId: 368636349