Remove enable_modules from ObjcCppSemantics
This was added as a workaround for an include validation problem on
Apple when the objc_library enable_modules attribute is set to true.
There were absolute paths to internal stl wrapper headers that were
showing up in .d files.  I verified that the absolute paths are no
longer showing up, and removing this work around didn't break any
builds.

This finally makes ObjcCppSemantics stateless which is the way it
should be!

PiperOrigin-RevId: 368850814