Make --experimental_prefer_mutual_xcode choose the local default version if it's mutually available.
Before, passing --experimental_prefer_mutual_xcode caused the local default to be ignored in favor of the newest mutually available Xcode version, even if the local default was available remotely as well.

RELNOTES: If --experimental_prefer_mutual_xcode is passed, Bazel will choose the local default (instead of the newest mutually available version) if it's available both locally and remotely.
PiperOrigin-RevId: 369501097