Allow rules in _builtins to use transitions, regardless of the allowlist
The allow list cannot reference symbols in @_builtins, so just always
allow them. Because _builtins is meant to be used for migrating existing
rules to Starlark, this should not represent an increase in the use of
transitions, so the lack of allowlisting here should not be significant.

The final migration out of builtins will require the allow list to be
filled appropriately.

PiperOrigin-RevId: 369656645