Include NSOS nodes in rewinding analysis of lost inputs from input discovery
The previous assumption, that lost inputs observed during input
discovery are either direct deps of the action execution node (or would
be, after input discovery is done), is not always true when NSOS is
enabled. Instead, they may be included in directly depended-on nested set
nodes.

This change ensures that any nested set node direct deps are considered
during action rewinding's analysis. These nodes may need to be
invalidated, to ensure that ArtifactNestedSetFunction's global artifact
value map is kept in sync with Skyframe.

PiperOrigin-RevId: 370107329