Change JavaPluginInfo to Starlark provider.
JavaPluginInfo needs to extend NativeInfo to expose it to Starlark and can no longer extend from TransitiveInfoProviders (there are several runtime checks preventing the use of wrong calls).

No exposure of the provider is done yet.

I had a more elaborate change prepared, which duplicated delegators to TransitiveInfoProviderMap in the JavaInfo class (fetchProvidersFromList, addProvider, getProvider, getProvidersFromTargetList).
However, I foresee only JavaPluginInfo to be both in JavaInfo and exposed independently to Starlark. So I decided for a smaller change, handling only JavaPluginInfo in a special way.

PiperOrigin-RevId: 370841099