Extend Skydoc with more general faking.
Unknown symbols are collected in the compile phase.

In the evaluation phase every unknown symbol returns a fake Starlark "super polymorphic" structure with any field and that is also callable. For example any expression like `xxx_module.func().field` won't throw an exception.

Possible problem is that the fake structure is not a list or a map. But I didn't find any functions that would be used during loading only that returned lists or maps and weren't already faked in a different way.

The faked structure also extends FakeProviderApi and has a name that equals to by what it was called. This correctly handles all the deleted providers. (except "objc" becomes "Objc"; it was incorrectly faked before).

PiperOrigin-RevId: 370849147