Remove FakeApi from Skydoc.
After a more general faking was done in unknown commit, there is no need to define all specific fake API.

This change is intended not to break fidelity of any mocking that we were already doing, but rather to summarize a ton of stubs by replacing them with a polymorphic object that could provide a superset of those behaviors. (It does "break" the ability of Stardoc to properly fail for some malformed programs, but we don't care about that use case, at least not until after all mocking is eliminated.) To the extent any existing user is broken due to an unintended change in mocking, we can consider forward fixing by replacing the generic polymorphic mock with something specially crafted for their use case.

PiperOrigin-RevId: 370853259