Stop trying to aggregate errors in action execution to include missing mandatory inputs. Once an exception is thrown, the global artifact key -> value map is in a potentially bad state: it might have stale data from an artifact that belonged to a nested set whose evaluations was aborted by an error. Note that the artifact itself may not be in error, but it may not have been evaluated yet, so the stale value is still there.
All of the approaches I considered (doing an additional pass over the nested set to remove undone keys after seeing an error, removing keys before re-evaluating them) seemed to either not work or open up dangerous races with other threads (in the keep-going case). Seemed not worth it for the slightly better error messaging.

I suspect that when we get rid of the non-nested set codepath, we'll be able to simplify the error handling into disjoint categories: missing mandatory artifacts, and IO exceptions/action execution exceptions.

PiperOrigin-RevId: 370914021