test command now adds 'default' output group if it is absent.
The --output_groups option determines which output groups are built for the
targets and aspects specified by command-line patterns, and artifacts in
other output groups do not affect the completion of those targets/aspects.

When testing is requested, the targets' executables in the 'default' output
group will be built even if the 'default' output group was not requested. The
executables will be built after the targets have reported their status however,
and if building the executable fails, the status will not be reported
correctly.

Now, 'test' commands will always include the 'default' output group during
the build phase, and failure to build the test executable will lead to the
target reporting a failure to build.

PiperOrigin-RevId: 371771843