Remove `defaultBuildOptions` from `PlatformMappingValue`.
It really only needs a list of options classes for parsing. Additionally, instead of passing it in, just store this list as a field. There shouldn't be any penalty for doing this, as equality checks will be fast since it's always the same instance in practice, and we don't serialize `PlatformMappingValue`.

After this, we can tear out `defaultBuildOptions` entirely. `PlatformMappingValue` was the last one using it.

PiperOrigin-RevId: 372134563