Simplify macOS arch setup
Quite a few settings only differ based on the architecture passed. This
hoists the architecture decision out of each individual action. There
are more places that can benefit from this but I only changed 1 here to
reduce the churn. This adds the `apply_simulator_compiler_flags` to more
configurations, but that flag is a no-op for non-simulator
configurations, so we don't need to branch on it in both places.