Accept JavaPluginInfo as plugins and exported_plugins parameter of java_common.compile and add exported_plugins parameter to JavaInfo constructor
java_common.compile now accepts either a list of JavaInfo or a list of JavaPluginInfo. A flag --incompatible_require_javaplugininfo_in_javacommon is added that disables supports for lists of JavaInfo, and at the same time also stops returning it from java_plugin rule.

JavaInfo constructor was missing exported_plugins parameter and it is added for consistency. Because it's new there is no support for a list of JavaInfo.

PiperOrigin-RevId: 373301285