Remove unneeded and conservative checking for LLVM compiler
The implicit ThinLTO features were only being added if we detected that this
was an LLVM compiler. This checking relied on the compiler toolchain including
the substring "llvm" which it may not always have when using LLVM. If a
specific compiler is needed this should be transitioned into the feature
configurations.

The easiest way to test this was to remove "llvm" from the mock toolchain name. With that change and without this fix the 3 implicit ThinLTO enablement tests in CcBinaryThinLtoTest.java fail, but with this fix they don't.

PiperOrigin-RevId: 373348337