Change the signature of tree artifact expansion to allow passing a reference to children from `TreeArtifactValue` to `ArtifactExpanderImpl`.
We currently create a copy of the set of children for each of the tree artifacts
when collecting input mappings. That can easily be avoided if we relax the
signature to allow subclasses of `Artifact`. Change the signature so that we
can take a reference to an already existing set of children in
`TreeArtifactValue` rather than copying it.

PiperOrigin-RevId: 373881983