Don't require that tests output to bin (instead of genfiles) for coverage
getInstrumentedFileManifest asserts `ruleContext.getRule().hasBinaryOutput()`. However, that doesn't assert that the target has an executable output. Rather, it asserts that the target's output goes to bin instead of genfiles. That's the preferred convention, but if a test sets `output_to_genfiles` for some reason (even though it probably shouldn't), having that return `InstrumentedFilesInfo` will cause it to suddenly start crashing in coverage mode. Which is pretty confusing.

RELNOTES: None.
PiperOrigin-RevId: 374180811