Rename Rule/RuleClass.hasBinaryOutput to outputsToBindir
This is consistent with the naming of genrule.output_to_bindir (and the output_to_genfiles parameter of rule). hasBinaryOutput is also a confusing name because hasBinaryOutput can be true even if a target does not have an executable output, or even if it does not have _any_ output. This is commonly the case since Starlark-defined rules output to bin by default.

RELNOTES: None.
PiperOrigin-RevId: 374424142