Update `apple_common.link_multi_arch_binary` to gate legacy functionality on `should_lipo = True` (the current default).
This change will allow rules_apple to migrate away from the Java code handling the lipo'ing of universal binaries and other hardcoded assumptions like the `binary_type` and `bundle_loader` attributes. When `should_lipo = False`, those attributes are not assumed to be present on the rule, and it is the calling rule's responsibility to pass the relevant linker flags/inputs into `link_multi_arch_binary` to support dylibs/loadable bundles.

The fields of the `struct` returned by `link_multi_arch_binary` change as follows. A new field is added—`outputs`—which is a list of `struct`s, one element per child configuration. Each element contains the following fields; the first three describe the configuration that was used to link the artifacts in the following four fields:

*   `platform`: a string denoting the target platform (`ios`, `macos`, `tvos`, `watchos`)
*   `architecture`: a string denoting the target CPU architecture (e.g., `x86_64`)
*   `environment`: a string denoting the target environment (`device`, `simulator`, `macabi`)
*   `binary`: the linked single architecture binary
*   `bitcode_symbols`: the single architecture Bitcode symbol map (or `None` if not requested)
*   `dsym_binary`: the single architecture dSYM binary (or `None` if not requested)
*   `linkmap`: the single architecture linkmap (or `None` if not requested)

Furthermore, when `should_lipo = False`, the multi-architecture `binary` and `binary_provider` fields of the top-level result `struct` are **suppressed.**

RELNOTES: None.
PiperOrigin-RevId: 374886744